Revert "feat(policy): extract PolicyService from policies router"
